# Conflicts:
#	app/src/main/java/com/gausslab/managerapp/checkinworkerbysite/CheckedInWorkersBySiteRecyclerViewAdapter.java
#	app/src/main/java/com/gausslab/managerapp/model/User.java

+++ app/src/main/java/com/gausslab/managerapp/workerinformation/WorkerInformationViewModel.java
@@ -17,50 +17,66 @@ import com.gausslab.managerapp.repository.AccidentRepository;
 import com.gausslab.managerapp.repository.UserRepository;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
-public class WorkerInformationViewModel extends ViewModel {
+public class WorkerInformationViewModel extends ViewModel
+{
     private final UserRepository userRepository = UserRepository.getInstance();
     private final AccidentRepository accidentRepository = AccidentRepository.getInstance();
 
     private final MutableLiveData<Boolean> userInformationLoaded = new MutableLiveData<>(false);
     private final MutableLiveData<List<AccidentHistory>> accidentHistoryList = new MutableLiveData<>();
     private final MutableLiveData<Event<Boolean>> addAccidentHistorySuccess = new MutableLiveData<Event<Boolean>>();
-
     private User currUser;
     private Drawable userImage;
-
-    private AccidentHistory accidentHistory;
-    private final MutableLiveData<Boolean> accidentHistoryLoaded = new MutableLiveData<>(false);
-
+    private AccidentHistory currEditingAccidentHistory;
     private List<AccidentHistory> deletedAccidentHistoryList = new ArrayList<>();
     private List<AccidentHistory> addedAccidentHistoryList = new ArrayList<>();
     private List<AccidentHistory> changedAccidentHistoryList = new ArrayList<>();
 
-    public void saveUser() {
-        userRepository.addOrUpdateUser(currUser, new CompletedCallback<Result<String>>() {
+    private Map<String, AccidentHistory> accidentHistoryMap = new HashMap<>();
+
+    public void saveUser()
+    {
+        userRepository.addOrUpdateUser(currUser, new CompletedCallback<Result<String>>()
+        {
             @Override
-            public void onComplete(Result<String> result) {
-                for (AccidentHistory newAdd : addedAccidentHistoryList) {
-                    accidentRepository.addAccidentHistory(newAdd, new CompletedCallback<Result<String>>() {
+            public void onComplete(Result<String> result)
+            {
+                for (AccidentHistory newAdd : addedAccidentHistoryList)
+                {
+                    accidentRepository.addAccidentHistory(newAdd, new CompletedCallback<Result<String>>()
+                    {
                         @Override
-                        public void onComplete(Result<String> result) {
+                        public void onComplete(Result<String> result)
+                        {
 
                         }
                     });
                 }
-                for (AccidentHistory del : deletedAccidentHistoryList) {
-                    accidentRepository.deleteAccidentHistory(del.getKeyValue(), new CompletedCallback<Result<String>>() {
-                        @Override
-                        public void onComplete(Result<String> result) {
+                for (AccidentHistory del : deletedAccidentHistoryList)
+                {
+                    if (del.getKeyValue() != null)
+                    {
+                        accidentRepository.deleteAccidentHistory(del.getKeyValue(), new CompletedCallback<Result<String>>()
+                        {
+                            @Override
+                            public void onComplete(Result<String> result)
+                            {
 
-                        }
-                    });
+                            }
+                        });
+                    }
                 }
-                for(AccidentHistory changed :changedAccidentHistoryList){
-                    accidentRepository.changeAccidentHistory(changed, new CompletedCallback<Result<String>>() {
+                for (AccidentHistory changed : changedAccidentHistoryList)
+                {
+                    accidentRepository.changeAccidentHistory(changed, new CompletedCallback<Result<String>>()
+                    {
                         @Override
-                        public void onComplete(Result<String> result) {
+                        public void onComplete(Result<String> result)
+                        {
 
                         }
                     });
@@ -72,7 +88,8 @@ public class WorkerInformationViewModel extends ViewModel {
         });
     }
 
-    public void addAccidentHistory(AccidentHistory toAdd) {
+    public void addAccidentHistory(AccidentHistory toAdd)
+    {
         List<AccidentHistory> updatedList = new ArrayList<>(accidentHistoryList.getValue());
         updatedList.add(toAdd);
         addedAccidentHistoryList.add(toAdd);
@@ -80,49 +97,58 @@ public class WorkerInformationViewModel extends ViewModel {
         addAccidentHistorySuccess.postValue(new Event<>(true));
     }
 
-    public void updateMemoText(String newMemoText) {
+    public void updateMemoText(String newMemoText)
+    {
         currUser.setMemo(newMemoText);
     }
 
-    public void deleteAccidentHistory(AccidentHistory toRemove) {
+    public void deleteAccidentHistory(AccidentHistory toRemove)
+    {
         List<AccidentHistory> updatedList = new ArrayList<>(accidentHistoryList.getValue());
         updatedList.remove(toRemove);
         deletedAccidentHistoryList.add(toRemove);
+        addedAccidentHistoryList.remove(toRemove);
         accidentHistoryList.postValue(updatedList);
     }
 
-    public void changeAccidentHistory(String description, String place, String date, String time) {
-        if (accidentHistory.getDescription().equals(description) &&
-                accidentHistory.getPlace().equals(place) &&
-                accidentHistory.getDate().equals(date) &&
-                accidentHistory.getTime().equals(time)){
+    public void changeAccidentHistory(String description, String place, String date, String time)
+    {
+        if (currEditingAccidentHistory.getDescription().equals(description) &&
+            currEditingAccidentHistory.getPlace().equals(place) &&
+            currEditingAccidentHistory.getDate().equals(date) &&
+            currEditingAccidentHistory.getTime().equals(time))
+        {
             addAccidentHistorySuccess.postValue(new Event<>(true));
             return;
         }
-        AccidentHistory newAccidentHistory = accidentHistory;
+
+        //boolean result = accidentHistoryList.getValue().remove(currEditingAccidentHistory);
+        //Log.d("DEBUG", "REMOVE RESULT - " + result);
+
+        currEditingAccidentHistory.setDescription(description);
+        currEditingAccidentHistory.setPlace(place);
+        currEditingAccidentHistory.setDate(date);
+        currEditingAccidentHistory.setTime(time);
+
         List<AccidentHistory> updatedList = new ArrayList<>(accidentHistoryList.getValue());
-        updatedList.remove(newAccidentHistory);
 
-        accidentHistory.setDescription(description);
-        accidentHistory.setPlace(place);
-        accidentHistory.setDate(date);
-        accidentHistory.setTime(time);
-
-
-        updatedList.add(accidentHistory);
-        changedAccidentHistoryList.add(accidentHistory);
-        accidentHistoryList.postValue(updatedList);
+        changedAccidentHistoryList.add(currEditingAccidentHistory);
+        accidentHistoryList.setValue(updatedList);
         addAccidentHistorySuccess.postValue(new Event<>(true));
     }
 
-    public void loadAllUserInformation(String phoneNumberOrUserName, boolean userHasPhoneNumber) {
+    public void loadAllUserInformation(String phoneNumberOrUserName, boolean userHasPhoneNumber)
+    {
         //로딩 안해도되는 조건
-        if (currUser != null) {
+        if (currUser != null)
+        {
             if (userHasPhoneNumber && currUser.getPhoneNumber().equals(phoneNumberOrUserName))
                 return;
             else if (!userHasPhoneNumber && currUser.getUserName().equals(phoneNumberOrUserName))
                 return;
         }
+
+        Log.d("DEBUG", "----loadAllUserInformation----");
 
         if (userHasPhoneNumber)
             loadUserInformation(phoneNumberOrUserName);
@@ -132,33 +158,41 @@ public class WorkerInformationViewModel extends ViewModel {
         loadAccidentHistoryListByUser(phoneNumberOrUserName);
     }
 
-    public LiveData<List<AccidentHistory>> getAccidentHistoryList() {
-        if (accidentHistoryList.getValue() == null) {
+    public LiveData<List<AccidentHistory>> getAccidentHistoryList()
+    {
+        if (accidentHistoryList.getValue() == null)
+        {
             accidentHistoryList.postValue(new ArrayList<>());
         }
         return accidentHistoryList;
     }
 
-    public Drawable getUserImage() {
+    public Drawable getUserImage()
+    {
         return userImage;
     }
 
-    public User getCurrUser() {
+    public User getCurrUser()
+    {
         return currUser;
     }
 
-    public LiveData<Boolean> isUserInformationLoaded() {
+    public LiveData<Boolean> isUserInformationLoaded()
+    {
         return userInformationLoaded;
     }
 
-    public LiveData<Event<Boolean>> isAddAccidentHistorySuccess() {
+    public LiveData<Event<Boolean>> isAddAccidentHistorySuccess()
+    {
         return addAccidentHistorySuccess;
     }
 
-    private void loadUserInformation(String phoneNumber) {
+    private void loadUserInformation(String phoneNumber)
+    {
         userRepository.getUserByPhoneNumber(phoneNumber, result ->
         {
-            if (result instanceof Result.Success) {
+            if (result instanceof Result.Success)
+            {
                 currUser = ((Result.Success<User>) result).getData();
 
                 userInformationLoaded.setValue(true);
@@ -166,58 +200,70 @@ public class WorkerInformationViewModel extends ViewModel {
         });
     }
 
-    private void loadNoPhoneNumberUserInformation(String userName) {
+    private void loadNoPhoneNumberUserInformation(String userName)
+    {
         userRepository.noPhoneNumberGetUser(userName, result ->
         {
-            if (result instanceof Result.Success) {
+            if (result instanceof Result.Success)
+            {
                 currUser = ((Result.Success<User>) result).getData();
                 userInformationLoaded.setValue(true);
             }
         });
     }
 
-    private void loadUserImage(String phoneNumber) {
-        userRepository.loadUserImageDrawable(phoneNumber, new CompletedCallback<Result<Drawable>>() {
+    private void loadUserImage(String phoneNumber)
+    {
+        userRepository.loadUserImageDrawable(phoneNumber, new CompletedCallback<Result<Drawable>>()
+        {
             @Override
-            public void onComplete(Result<Drawable> drawableResult) {
-                if (drawableResult instanceof Result.Success) {
+            public void onComplete(Result<Drawable> drawableResult)
+            {
+                if (drawableResult instanceof Result.Success)
+                {
                     userImage = ((Result.Success<Drawable>) drawableResult).getData();
-                } else {
+                }
+                else
+                {
 
                 }
             }
         });
     }
 
-    private void loadAccidentHistoryListByUser(String phoneNumber) {
-        accidentRepository.registerAccidentHistoryListListener(phoneNumber, new ListenerCallback<List<AccidentHistory>>() {
+    private void loadAccidentHistoryListByUser(String phoneNumber)
+    {
+        accidentRepository.registerAccidentHistoryListListener(phoneNumber, new ListenerCallback<List<AccidentHistory>>()
+        {
             @Override
-            public void onUpdate(List<AccidentHistory> result) {
+            public void onUpdate(List<AccidentHistory> result)
+            {
+                for (AccidentHistory ah : result)
+                {
+                    accidentHistoryMap.put(ah.getKeyValue(), ah);
+                }
                 accidentHistoryList.postValue(result);
             }
         });
     }
 
-    public void loadAccidentHistory(String key) {
-        accidentRepository.getAccidentHistory(key, new CompletedCallback<Result<AccidentHistory>>() {
-            @Override
-            public void onComplete(Result<AccidentHistory> result) {
-                if (result instanceof Result.Success) {
-                    accidentHistory = ((Result.Success<AccidentHistory>) result).getData();
-                    accidentHistoryLoaded.postValue(true);
-                } else {
-                    //Error
-                }
-            }
-        });
+    public void setCurrEditingAccidentHistoryByKey(String key)
+    {
+        currEditingAccidentHistory = getAccidentHistoryByKey(key);
     }
 
-    public LiveData<Boolean> isAccidentHistoryLoaded() {
-        return accidentHistoryLoaded;
+    private AccidentHistory getAccidentHistoryByKey(String key)
+    {
+        if (accidentHistoryMap.containsKey(key))
+        {
+            return accidentHistoryMap.get(key);
+        }
+        return null;
     }
 
-    public AccidentHistory getAccidentHistory() {
-        return accidentHistory;
+    public AccidentHistory getCurrEditingAccidentHistory()
+    {
+        return currEditingAccidentHistory;
     }
 
 }
